rename rpc::Client -> ServerClient
