+++ src/common/rpc/mod.rs
@@ -21,12 +21,12 @@
 
 pub(crate) mod builder;
 pub(crate) mod channel;
-pub(crate) mod client;
 pub(crate) mod cluster_client;
+pub(crate) mod core_client;
 pub(crate) mod server_client;
 pub(crate) mod transaction;
 
 pub(crate) use channel::Channel;
-pub(crate) use client::Client;
 pub(crate) use cluster_client::{ClusterClient, ClusterClientManager};
+pub(crate) use core_client::CoreClient;
 pub(crate) use server_client::ServerClient;
